refactor(artifacts): migrate consumers to canonical modules
