+++ surfsense_web/features/artifacts-library/ui/library-image-viewer.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { Image, ImageLoading } from "@/components/tool-ui/image";
-import { useArtifactImage } from "@/features/artifacts/use-artifact-image";
+import { useArtifactImage } from "@/features/artifacts/hooks/use-artifact-image";
 
 export function LibraryImageViewer({
 	artifactId,
