+++ surfsense_web/features/chat-artifacts/ui/artifact-row.tsx
@@ -1,11 +1,11 @@
 import { useAtomValue, useSetAtom } from "jotai";
-import { openArtifactPanelAtom } from "@/atoms/chat/artifact-panel.atom";
 import { activeWorkspaceIdAtom } from "@/atoms/workspaces/workspace-query.atoms";
 import { TextShimmerLoader } from "@/components/prompt-kit/loader";
-import { ArtifactDownloadButton } from "@/features/artifacts/artifact-download-button";
-import { ArtifactFormatIcon } from "@/features/artifacts/artifact-format-icon";
-import { ArtifactFormatLabel } from "@/features/artifacts/artifact-format-label";
-import { artifactDownloadPath } from "@/features/artifacts/download-file";
+import { artifactDownloadPath } from "@/features/artifacts/api/artifact-download-path";
+import { openArtifactPanelAtom } from "@/features/artifacts/state/artifact-panel.atom";
+import { ArtifactDownloadButton } from "@/features/artifacts/ui/artifact-download-button";
+import { ArtifactFormatIcon } from "@/features/artifacts/ui/artifact-format-icon";
+import { ArtifactFormatLabel } from "@/features/artifacts/ui/artifact-format-label";
 import { useMediaQuery } from "@/hooks/use-media-query";
 import { openChatArtifact } from "../lib/open-chat-artifact";
 import type { ChatArtifact } from "../model/artifact";
